Combine WorldMapArea tracking into singular WorldMapAreaManager
